Change PeerSessionInfos log message to debug level
We are no longer including the peer client information in each BlockChainSync log message (since it bloats the logs). However, this information can be useful for debugging purposes and the initial peering could have happened long in the past, so we promote the PeerSessionInfos log message from trace to debug verbosity so that this information is more available in logs for debugging purposes.